+++ tools/calc_xyz_transform.py
@@ -9,6 +9,7 @@ from coloraide import algebra as alg  # noqa: E402
 
 white_d65 = util.xy_to_xyz((0.31270, 0.32900))
 white_d50 = util.xy_to_xyz((0.34570, 0.35850))
+white_aces = util.xy_to_xyz((0.32168, 0.33767))
 
 
 def pprint(value):
@@ -42,6 +43,12 @@ def get_matrix(wp, space):
     elif space == 'prophoto-rgb':
         x = [0.7347, 0.1596, 0.0366]
         y = [0.2653, 0.8404, 0.0001]
+    elif space == 'aces-ap0':
+        x = [0.7347, 0.0, 0.0001]
+        y = [0.2653, 1.0, -0.0770]
+    elif space == 'aces-ap1':
+        x = [0.713, 0.165, 0.128]
+        y = [0.293, 0.830, 0.044]
     else:
         raise ValueError
 
@@ -85,6 +92,20 @@ if __name__ == "__main__":
 
     print('===== ProPhoto =====')
     to_xyz, from_xyz = get_matrix(white_d50, 'prophoto-rgb')
+    print('--- rgb -> xyz ---')
+    pprint(to_xyz)
+    print('--- xyz -> rgb ---')
+    pprint(from_xyz)
+
+    print('===== ACES =====')
+    to_xyz, from_xyz = get_matrix(white_aces, 'aces-ap0')
+    print('--- rgb -> xyz ---')
+    pprint(to_xyz)
+    print('--- xyz -> rgb ---')
+    pprint(from_xyz)
+
+    print('===== ACEScc =====')
+    to_xyz, from_xyz = get_matrix(white_aces, 'aces-ap1')
     print('--- rgb -> xyz ---')
     pprint(to_xyz)
     print('--- xyz -> rgb ---')