+++ docs/src/markdown/about/changelog.md
@@ -3,6 +3,7 @@
 ## 1.0b2
 
 - **NEW**: Added HPLuv space described in the HSLuv spec.
+- **NEW**: Added new color spaces: ACES 2065-1, ACEScg, ACEScc, and ACEScct.
 - **NEW**: Contrast is now exposed as a plugin to allow for future expansion of approaches. While there is currently
   only one approach, methods can be selected via the `method` attribute.